+++ singer/schema_generation.py
@@ -37,13 +37,6 @@ def add_observations(acc, path, data):
             return acc
         except (ValueError, TypeError):
             pass
-        try:
-            # If the string parses as a date, add an observation that it's a date
-            dateutil.parser.parse(data)
-            add_observation(acc, path + ["date"])
-            return acc
-        except (dateutil.parser.ParserError, OverflowError):
-            pass
         add_observation(acc, path + ["string"])
     elif isinstance(data, bool):
         add_observation(acc, path + ["boolean"])
@@ -60,7 +53,10 @@ def add_observations(acc, path, data):
 
 def to_json_schema(obs):
     types = []
-    for key in obs:
+    # add schema types in a specific order to anyOf list
+    for key in ['array', 'object', 'number', 'integer', 'boolean', 'string', 'null']:
+        if key not in obs:
+            continue
 
         result = {'type': ['null']}
 
@@ -74,10 +70,6 @@ def to_json_schema(obs):
         elif key == 'array':
             result['type'] += ['array']
             result['items'] = to_json_schema(obs['array'])
-
-        elif key == 'date':
-            result['type'] += ['string']
-            result['format'] = 'date-time'
 
         elif key == 'string':
             result['type'] += ['string']